refactor: apply hexagonal architecture to project structure
- Renamed PriceService to PriceUseCase and moved it to domain.port.in
- Moved PriceRepository interface to domain.port.out
- Created H2 adapter for PriceRepository in infrastructure.adapter.out.h2
- Moved PriceController to infrastructure.adapter.in.rest
- Moved DTO to shared.dto
- Updated package structure to reflect hexagonal architecture principles
